+++ include/core/utility/UtilIO.hpp
@@ -37,7 +37,7 @@ namespace util {
 *  \param[out] transformation matrix
 *  \date       2017-04-07
 */
-static Matrix4 readT(const std::string &path, const float scale = 1) {
+static Matrix4 readT(const std::string &path, const double scale = 1) {
     Matrix4 T;
     if (path.substr(path.find_last_of(".")) != ".dat") {
         Logging::error("Wrong file type", "UtilIO");
@@ -47,13 +47,13 @@ static Matrix4 readT(const std::string &path, const float scale = 1) {
     if (fin.is_open()) {
         for (int row = 0; row < 4; ++row)
             for (int col = 0; col < 4; ++col) {
-                float item = 0.0;
+                double item = 0.0;
                 fin >> item;
                 T(row, col) = item;
             }
         fin.close();
     }
-    T.block<3,1>(0,3) = T.block<3,1>(0,3) * scale;
+    T.block<3, 1>(0, 3) = T.block<3, 1>(0, 3) * scale;
     return T;
 }
