+++ include/environment/robot/Joint.h
@@ -33,18 +33,18 @@ namespace ippp {
 class Joint {
   public:
     Joint();
-    Joint(float minBound, float maxBound);
-    Joint(float minBound, float maxBound, std::shared_ptr<ModelContainer> model);
+    Joint(double minBound, double maxBound);
+    Joint(double minBound, double maxBound, std::shared_ptr<ModelContainer> model);
 
     void setModel(std::shared_ptr<ModelContainer> &model);
     std::shared_ptr<ModelContainer> getModel();
 
-    void setBoundaries(float minBound, float maxBound);
-    void getBoundaries(float &minBound, float &maxBound);
+    void setBoundaries(double minBound, double maxBound);
+    void getBoundaries(double &minBound, double &maxBound);
 
   private:
-    float m_minBound = 0;
-    float m_maxBound = 0;
+    double m_minBound = 0;
+    double m_maxBound = 0;
     std::shared_ptr<ModelContainer> m_model = nullptr;
 };
