+++ source/core/collisionDetection/CollisionDetection2D.cpp
@@ -86,7 +86,7 @@ bool CollisionDetection2D::controlTrajectory(std::vector<Vector2> &vecs) {
 *  \param[out] binary result of collision (true if in collision)
 *  \date       2016-06-30
 */
-bool CollisionDetection2D::checkPoint2D(float x, float y) {
+bool CollisionDetection2D::checkPoint2D(double x, double y) {
     if (m_minBoundary[0] >= x || x >= m_maxBoundary[0] || m_minBoundary[1] >= y || y >= m_maxBoundary[1]) {
         Logging::debug("Point out of workspace", this);
         return true;
